requirements.txt
Django==1.7.7
djangorestframework==3.1.1
drf-extensions==0.2.7
wsgiref==0.1.2
